+++ src/core/utils.h
@@ -1,4 +1,0 @@
-#pragma once
-#ifndef UTILS_H
-#define UTILS_H
-#endif // !UTILS_H